feat(combobox): intégration dans formulaire LEE fonctionnel manque les tests
